pyproject.toml
[project]
name = "gian-website-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.128.0",
  "pydantic>=2.12.5",
  "pydantic-ai-slim[openai]>=1.44.0",
  "pydantic-settings>=2.12.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
autoSearchPaths = true
autoImportCompletions = true

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
